+++ packages/nextra/src/normalize-pages.ts
@@ -1,5 +1,6 @@
-import * as z from 'zod' // use `import * as z` for better tree-shaking https://github.com/colinhacks/zod/issues/2596#issuecomment-1643053289
+import type { z } from 'zod'
 import { ERROR_ROUTES } from './constants'
+import type { displaySchema, menuItemSchema, pageThemeSchema } from './schemas'
 import type { Folder, MdxFile, PageMapItem } from './types'
 
 const DEFAULT_PAGE_THEME: PageTheme = {
@@ -15,65 +16,10 @@ const DEFAULT_PAGE_THEME: PageTheme = {
   typesetting: 'default'
 }
 
-const pageThemeSchema = z.strictObject({
-  breadcrumb: z.boolean(),
-  collapsed: z.boolean(),
-  footer: z.boolean(),
-  layout: z.enum(['default', 'full', 'raw']),
-  navbar: z.boolean(),
-  pagination: z.boolean(),
-  sidebar: z.boolean(),
-  timestamp: z.boolean(),
-  toc: z.boolean(),
-  typesetting: z.enum(['default', 'article'])
-})
-
 export type PageTheme = z.infer<typeof pageThemeSchema>
-
-/**
- * An option to control how an item should be displayed in the sidebar:
- * - `normal`: the default behavior, item will be displayed
- * - `hidden`: the item will not be displayed in the sidebar entirely
- * - `children`: if the item is a folder, itself will be hidden but all its children will still be processed
- */
-const displaySchema = z.enum(['normal', 'hidden', 'children'])
-const titleSchema = z.string()
-
-const linkItemSchema = z.strictObject({
-  href: z.string(),
-  newWindow: z.boolean(),
-  title: titleSchema
-})
-
-const menuItemSchema = z.strictObject({
-  display: displaySchema.optional(),
-  items: z.record(linkItemSchema.partial({ href: true, newWindow: true })),
-  title: titleSchema,
-  type: z.literal('menu')
-})
-
-const separatorItemSchema = z.strictObject({
-  title: titleSchema,
-  type: z.literal('separator')
-})
-
-const itemSchema = linkItemSchema
-  .extend({
-    display: displaySchema,
-    theme: pageThemeSchema,
-    title: titleSchema,
-    type: z.enum(['page', 'doc'])
-  })
-  .deepPartial()
 
 type Display = z.infer<typeof displaySchema>
 type IMenuItem = z.infer<typeof menuItemSchema>
-
-export const metaSchema = z
-  .string()
-  .or(menuItemSchema)
-  .or(separatorItemSchema)
-  .or(itemSchema)
 
 function extendMeta(
   meta: string | Record<string, any> = {},
